fix(admin): pagesのリファクタ
